+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the "jsmm" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+## [0.2.0] - 2019-11-10
+- jsmm.addChild command (active with jsmm view)
+- jsmm.addSibling command (active with jsmm view)
+- jsmm.remove command (active with jsmm view)
+- js mind map command category
+- default keybinding
 
 ## [0.1.0] - 2019-11-03
 - Initial release